+++ poppy/optics.py
@@ -21,7 +21,7 @@ __all__ = ['AnalyticOpticalElement', 'ScalarTransmission', 'InverseTransmission'
            'CircularOcculter', 'BarOcculter', 'FQPM_FFT_aligner',  'CircularAperture',
            'HexagonAperture', 'MultiHexagonAperture', 'NgonAperture', 'RectangleAperture',
            'SquareAperture', 'SecondaryObscuration', 'AsymmetricSecondaryObscuration',
-           'ThinLens', 'ZernikeAberration', 'ParameterizedAberration', 'CompoundAnalyticOptic']
+           'ThinLens', 'CompoundAnalyticOptic']
 
 # ------ Generic Analytic elements -----
 
@@ -1416,146 +1416,6 @@ class ThinLens(CircularAperture):
         lens_phasor = np.exp(1.j * 2 * np.pi * defocus_zernike * aperture_intensity)
 
         return lens_phasor
-
-
-class ParameterizedAberration(AnalyticOpticalElement):
-    """
-    Define an optical element in terms of its distortion as decomposed into a set or orthonormal
-    basis functions (e.g. Zernikes, Hexikes, etc.). Included basis functions are normalized
-    such that user-provided coefficients correspond to meters RMS wavefront aberration for that
-    basis function.
-
-    Parameters
-    ----------
-    coefficients : iterable of numbers
-        The contribution of each term to the final distortion, in meters RMS wavefront error.
-        The coefficients are interpreted as indices in the order of Noll et al. 1976: the first
-        term corresponds to j=1, second to j=2, and so on.
-    radius : float
-        Pupil radius, in meters. Defines the region of the input wavefront array over which
-        the distortion terms will be evaluated. For non-circular pupils, this should be the
-        circle circumscribing the actual pupil shape.
-    basis_factory : callable
-        basis_factory will be called with the arguments `nterms`, `rho`, and `theta`.
-        `nterms` specifies how many terms to compute, starting with the j=1 term in the
-        Noll indexing convention for `nterms` = 1 and counting up. `rho` and `theta` are square
-        arrays holding the rho and theta coordinates at each pixel in the pupil plane.
-
-        `rho` is normalized such that `rho` == 1.0 for pixels at `radius` meters from
-        the center.
-    """
-    def __init__(self, name="Parameterized Distortion", coefficients=None, radius=None,
-                 basis_factory=None, **kwargs):
-        if not isinstance(basis_factory, collections.Callable):
-            raise ValueError("'basis_factory' must be a callable that can "
-                             "calculate basis functions")
-        try:
-            self.radius = float(radius)
-        except TypeError:
-            raise ValueError("'radius' must be the radius of a circular aperture in meters"
-                             "(optionally circumscribing a pupil of another shape)")
-        self.coefficients = coefficients
-        self.basis_factory = basis_factory
-        AnalyticOpticalElement.__init__(self, name=name, planetype=_PUPIL, **kwargs)
-
-    def getPhasor(self, wave):
-        rho, theta = _wave_to_rho_theta(wave, self.radius)
-        combined_distortion = np.zeros(rho.shape)
-
-        nterms = len(self.coefficients)
-        computed_terms = self.basis_factory(nterms=nterms, rho=rho, theta=theta)
-
-        for idx, coefficient in enumerate(self.coefficients):
-            if coefficient == 0.0:
-                continue  # save the trouble of a multiply-and-add of zeros
-            combined_distortion += coefficient * computed_terms[idx]
-
-        opd_as_phase = 2 * np.pi * combined_distortion / wave.wavelength
-        return np.exp(1.0j * opd_as_phase)
-
-
-def _wave_to_rho_theta(wave, pupil_radius):
-    """
-    Return wave coordinates in (rho, theta) for a Wavefront object normalized such that
-    rho == 1.0 at the pupil radius
-
-    Parameters
-    ----------
-    wave : Wavefront
-        Wavefront object with a `coordinates` method that returns (y, x)
-        coordinate arrays in meters in the pupil plane
-    pupil_radius : float, optional
-        Radius (in meters) of a circle circumscribing the pupil.
-        If `None`, this function will attempt to guess the radius from the wave
-        intensity array using `_guess_pupil_radius`.
-    """
-    y, x = wave.coordinates()
-    r = np.sqrt(x ** 2 + y ** 2)
-
-    rho = r / pupil_radius
-    theta = np.arctan2(y / pupil_radius, x / pupil_radius)
-
-    return rho, theta
-
-
-class ZernikeAberration(CircularAperture):
-    """
-    Define an optical element in terms of its Zernike components by providing coefficients
-    for each Zernike term modeled by the analytic optical element.
-
-    Parameters
-    ----------
-    coefficients : iterable of 3-tuples
-        Each 3-tuple in coefficients must be of the form (n, m, k), where n and m are the integer
-        radial degree and azimuthal frequency indices of the Zernike, and k is the RMS wavefront
-        aberration over the pupil in meters for that Zernike component.
-    radius : float
-        Pupil radius, in meters, over which the Zernike terms should be computed such that
-        rho = 1 at r = `radius`.
-    """
-    def __init__(self, name="Zernike Optic", coefficients=None, radius=None, **kwargs):
-        try:
-            self.radius = float(radius)
-        except TypeError:
-            raise ValueError("'radius' must be the radius of a circular aperture in meters"
-                             "(optionally circumscribing a pupil of another shape)")
-
-        def _validate_coefficients():
-            if coefficients is None:
-                return False
-            for coeff_tuple in coefficients:
-                if len(coeff_tuple) != 3:
-                    return False
-                if not int(coeff_tuple[0]) == coeff_tuple[0] \
-                        or not int(coeff_tuple[1]) == coeff_tuple[1]:
-                    return False
-            return True
-
-        if not _validate_coefficients():
-            raise ValueError("Coefficients must be supplied as a sequence of tuples with (n, m, k) "
-                             "where n, m are the indices of the Zernike, and k is a leading "
-                             "coefficient in meters of wavefront error. "
-                             "e.g. coefficients=[(2, 0, 0.214), (2, -2, 0.02)]")
-        self.coefficients = coefficients
-        CircularAperture.__init__(self, name=name, radius=self.radius, **kwargs)
-
-    def getPhasor(self, wave):
-        rho, theta = _wave_to_rho_theta(wave, self.radius)
-
-        # the Zernike optic, being circular, is implicitly also a circular aperture:
-        aperture_intensity = CircularAperture.getPhasor(self, wave)
-
-        combined_zernikes = np.zeros(wave.shape, dtype=np.float64)
-        for n, m, k in self.coefficients:
-            combined_zernikes += k * zernike.zernike(n, m, rho=rho, theta=theta,
-                                                     mask_outside=True, outside=0.0)
-
-        combined_zernikes *= aperture_intensity
-
-        opd_as_phase = 2 * np.pi * combined_zernikes / wave.wavelength
-        lens_phasor = np.exp(1.j * opd_as_phase)
-        return lens_phasor
-
 
 #------ generic analytic optics ------
